backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
gunicorn==21.2.0
python-dotenv==1.0.0
openai>=1.0.0
requests>=2.28.0
cryptography>=42.0.0
Flask-SocketIO>=5.3.6
simple-websocket>=1.0.0
